+++ eve_elastic/__init__.py
@@ -3,5 +3,5 @@
 __version__ = '0.1.6'
 
 # flake8: noqa
-from .elastic import Elastic
+from .elastic import Elastic, get_es, get_indices
 from .validation import Validator